FEAT: 레벨2 계산기 기능 추가

+++ src/main/java/com/example/calculator2/AbstractOperation.java
@@ -1,5 +1,5 @@
 package com.example.calculator2;
 
 public abstract class AbstractOperation {
-    abstract double operate(int firstNumber, int secondNumber);
+    abstract double operate(long firstNumber, long secondNumber);
 }
