+++ src/main/java/com/example/calculator2/AddOperation.java
@@ -3,7 +3,7 @@ package com.example.calculator2;
 public class AddOperation extends AbstractOperation {
 
     @Override
-    double operate(int firstNumber, int secondNumber) {
+    double operate(long firstNumber, long secondNumber) {
         return firstNumber + secondNumber;
     }
 }
