+++ src/main/java/com/example/calculator2/DivideOperation.java
@@ -2,7 +2,7 @@ package com.example.calculator2;
 
 public class DivideOperation extends AbstractOperation {
     @Override
-    double operate(int firstNumber, int secondNumber) {
+    double operate(long firstNumber, long secondNumber) {
         if (secondNumber == 0) {
             System.out.println("Divide by zero");
             return 0;
