+++ src/main/java/com/example/calculator2/SubstractOperation.java
@@ -2,7 +2,7 @@ package com.example.calculator2;
 
 public class SubstractOperation extends AbstractOperation {
     @Override
-    double operate(int firstNumber, int secondNumber) {
+    double operate(long firstNumber, long secondNumber) {
         return firstNumber - secondNumber;
     }
 }
